Fix customization folder
