+++ src/components/Blocks/Slider/utils.js
