+++ src/customizations/@kitconcept/volto-slider-block/components/DefaultBody.jsx
@@ -1,143 +1,11 @@
-import React from 'react';
-import { useIntl, defineMessages } from 'react-intl';
-import { getTeaserImageURL } from '../helpers';
-import { flattenToAppURL } from '@plone/volto/helpers';
-import { Icon, MaybeWrap, UniversalLink } from '@plone/volto/components';
-import { Input, Button, Message } from 'semantic-ui-react';
-import cx from 'classnames';
-import navTreeSVG from '@plone/volto/icons/nav.svg';
-import imageBlockSVG from '@plone/volto/components/manage/Blocks/Image/block-image.svg';
-import config from '@plone/volto/registry';
+/**
+ * OVERRIDE DefaultBody.jsx
+ * REASON: Adding Button and flagAlign in slider block.
+ * FILE: https://github.com/kitconcept/volto-slider-block/blob/master/src/components/DefaultBody.jsx
+ * DATE: 2023-08-02
+ * DEVELOPER: @iRohitSingh
+ */
 
-const messages = defineMessages({
-  PleaseChooseContent: {
-    id: 'Please choose an existing content as source for this element',
-    defaultMessage:
-      'Please choose an existing content as source for this element',
-  },
-  moreInfo: {
-    id: 'moreInfo',
-    defaultMessage: 'More info',
-  },
-  source: {
-    id: 'Source',
-    defaultMessage: 'Source',
-  },
-  ButtonText: {
-    id: 'Continue reading',
-    defaultMessage: 'Continue reading',
-  },
-});
-
-const DefaultImage = (props) => <img {...props} alt={props.alt || ''} />;
-
-const SliderBody = ({
-  index,
-  onChangeBlock,
-  block,
-  data,
-  dataBlock,
-  isEditMode,
-  openObjectBrowser,
-}) => {
-  const intl = useIntl();
-  const href = data.href?.[0];
-  const image = data.preview_image?.[0];
-
-  const hasImageComponent = config.getComponent('Image').component;
-  const Image = config.getComponent('Image').component || DefaultImage;
-  const defaultImageSrc =
-    href && flattenToAppURL(getTeaserImageURL({ href, image }));
-
-  const handleClick = () => {
-    openObjectBrowser({
-      onSelectItem: (url, document) => {
-        dataBlock.slides[index].title = document.Title;
-        dataBlock.slides[index].description = document.Description;
-        dataBlock.slides[index].href = [
-          {
-            '@id': document['@id'],
-            Title: document.Title,
-            Description: document.Description,
-            title: document.Title,
-            image_field: document.image_field,
-            hasPreviewImage: document.hasPreviewImage,
-          },
-        ];
-        onChangeBlock(block, dataBlock);
-      },
-      mode: 'link',
-    });
-  };
-
-  return (
-    <div
-      className={cx('grid-teaser-item top', {
-        'empty-slide': !href && isEditMode,
-      })}
-    >
-      {!href && isEditMode && (
-        <Message>
-          <div className="grid-teaser-item default">
-            <img src={imageBlockSVG} alt="" />
-            <p>{intl.formatMessage(messages.PleaseChooseContent)}</p>
-            <div className="toolbar-inner">
-              <Button.Group>
-                <Button onClick={handleClick} icon basic>
-                  <Icon name={navTreeSVG} size="24px" />
-                </Button>
-              </Button.Group>
-              <Input
-                placeholder={`${intl.formatMessage(messages.source)}...`}
-                onClick={handleClick}
-                onFocus={(e) => e.target.blur()}
-              />
-            </div>
-          </div>
-        </Message>
-      )}
-      {href && (
-        <div className="teaser-item top">
-          <MaybeWrap
-            condition={!isEditMode}
-            as={UniversalLink}
-            href={href['@id']}
-            target={data.openLinkInNewTab ? '_blank' : null}
-            tabIndex="-1"
-          >
-            {(href?.hasPreviewImage || image) && (
-              <div className="highlight-image-wrapper gradient">
-                <Image
-                  src={hasImageComponent ? href : defaultImageSrc}
-                  alt=""
-                  loading="lazy"
-                />
-              </div>
-            )}
-            <div
-              className={cx(
-                'teaser-item-title fix-width-issue',
-                `has--slider--flagAlign--${data.flagAlign}`,
-              )}
-            >
-              <div className="title">
-                {data?.head_title && (
-                  <span className="supertitle">{data?.head_title}</span>
-                )}
-                <h2>{data?.nav_title || data?.title}</h2>
-              </div>
-              <p>{data?.description}</p>
-              {!data.hideButton && (
-                <Button tabIndex={'-1'}>
-                  {data.buttonText || intl.formatMessage(messages.ButtonText)}
-                </Button>
-              )}
-            </div>
-          </MaybeWrap>
-        </div>
-      )}
-    </div>
-  );
-};
+import SliderBody from '../../../../components/Blocks/Slider/DefaultBody';
 
 export default SliderBody;
