--- src/customizations/@kitconcept/volto-slider-block/helpers/index.js
@@ -1,2 +1,0 @@
-import { getTeaserImageURL } from './Image/Image';
-export { getTeaserImageURL };
